requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.10.6
python-multipart==0.0.20
python-dotenv==1.1.1
typing-extensions==4.15.0
anyio==3.7.1
httpx==0.24.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
black==25.9.0
flake8==7.3.0
isort==6.1.0
mypy==1.18.2
bandit==1.7.10
pre-commit==4.3.0
ipython==9.6.0